+++ services/roomService/src/requestHandlers/CoveyTownRequestHandlers.ts
@@ -4,6 +4,7 @@ import Player from '../types/Player';
 import { CoveyTownList, UserLocation } from '../CoveyTypes';
 import CoveyTownListener from '../types/CoveyTownListener';
 import CoveyTownsStore from '../lib/CoveyTownsStore';
+import {MessageData} from '../types/MessageData';
 
 /**
  * The format of a request to join a Town in Covey.Town, as dispatched by the server middleware
@@ -67,16 +68,6 @@ export interface TownDeleteRequest {
   coveyTownPassword: string;
 }
 
-export interface MessageData {
-  senderName: string;
-  senderID: string;
-  receiverName: string;
-  receiverID: string;
-  roomName: string;
-  roomID: string;
-  content: string;
-  time: string;
-}
 
 /**
  * Payload sent by the client to update a Town.
