+++ services/roomService/src/types/MessageData.ts
@@ -1,0 +1,10 @@
+export interface MessageData {
+  senderName: string;
+  senderID: string;
+  receiverName: string;
+  receiverID: string;
+  roomName: string;
+  roomID: string;
+  content: string;
+  time: string;
+}
